separate distributions util and hmc util
