+++ test/test_distributions_util.py
@@ -1,0 +1,69 @@
+import pytest
+import scipy.special as osp_special
+from numpy.testing import assert_allclose
+
+import jax.numpy as np
+from jax import grad, jit, lax
+from jax.util import partial
+
+from numpyro.distributions.util import xlog1py, xlogy
+
+_zeros = partial(lax.full_like, fill_value=0)
+
+
+@pytest.mark.parametrize('x, y', [
+    (np.array([1]), np.array([1, 2, 3])),
+    (np.array([0]), np.array([0, 0])),
+    (np.array([[0.], [0.]]), np.array([1., 2.])),
+])
+@pytest.mark.parametrize('jit_fn', [False, True])
+def test_xlogy(x, y, jit_fn):
+    fn = xlogy if not jit_fn else jit(xlogy)
+    assert_allclose(fn(x, y), osp_special.xlogy(x, y))
+
+
+@pytest.mark.parametrize('x, y, grad1, grad2', [
+    (np.array([1., 1., 1.]), np.array([1., 2., 3.]),
+     np.log(np.array([1, 2, 3])), np.array([1., 0.5, 1./3])),
+    (np.array([1.]), np.array([1., 2., 3.]),
+     np.sum(np.log(np.array([1, 2, 3]))), np.array([1., 0.5, 1./3])),
+    (np.array([1., 2., 3.]), np.array([2.]),
+     np.log(np.array([2., 2., 2.])), np.array([3.])),
+    (np.array([0.]), np.array([0, 0]),
+     np.array([-float('inf')]), np.array([0, 0])),
+    (np.array([[0.], [0.]]), np.array([1., 2.]),
+     np.array([[np.log(2.)], [np.log(2.)]]), np.array([0, 0])),
+])
+def test_xlogy_jac(x, y, grad1, grad2):
+    assert_allclose(grad(lambda x, y: np.sum(xlogy(x, y)))(x, y), grad1)
+    assert_allclose(grad(lambda x, y: np.sum(xlogy(x, y)), 1)(x, y), grad2)
+
+
+@pytest.mark.parametrize('x, y', [
+    (np.array([1]), np.array([0, 1, 2])),
+    (np.array([0]), np.array([-1, -1])),
+    (np.array([[0.], [0.]]), np.array([1., 2.])),
+])
+@pytest.mark.parametrize('jit_fn', [False, True])
+def test_xlog1py(x, y, jit_fn):
+    fn = xlog1py if not jit_fn else jit(xlog1py)
+    assert_allclose(fn(x, y), osp_special.xlog1py(x, y))
+
+
+@pytest.mark.parametrize('x, y, grad1, grad2', [
+    (np.array([1., 1., 1.]), np.array([0., 1., 2.]),
+     np.log(np.array([1, 2, 3])), np.array([1., 0.5, 1./3])),
+    (np.array([1., 1., 1.]), np.array([-1., 0., 1.]),
+     np.log(np.array([0, 1, 2])), np.array([float('inf'), 1., 0.5])),
+    (np.array([1.]), np.array([0., 1., 2.]),
+     np.sum(np.log(np.array([1, 2, 3]))), np.array([1., 0.5, 1./3])),
+    (np.array([1., 2., 3.]), np.array([1.]),
+     np.log(np.array([2., 2., 2.])), np.array([3.])),
+    (np.array([0.]), np.array([-1, -1]),
+     np.array([-float('inf')]), np.array([0, 0])),
+    (np.array([[0.], [0.]]), np.array([1., 2.]),
+     np.array([[np.log(6.)], [np.log(6.)]]), np.array([0, 0])),
+])
+def test_xlog1py_jac(x, y, grad1, grad2):
+    assert_allclose(grad(lambda x, y: np.sum(xlog1py(x, y)))(x, y), grad1)
+    assert_allclose(grad(lambda x, y: np.sum(xlog1py(x, y)), 1)(x, y), grad2)
